% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bnpgrangertest.R
\name{extract_chain_irf}
\alias{extract_chain_irf}
\title{Extract the full MCMC chain associated to the IRFs.}
\usage{
extract_chain_irf(object)
}
\arguments{
\item{object}{An instance of class \code{bnpgrangertest}.}
}
\value{
A tibble with the chain of the IRFs and the following variables:
\itemize{
  \item \code{iter}: The MCMC iteration.
  \item \code{horizon}: The IRF horizon.
  \item \code{cause_var}: The causing variable in the relationship.
  \item \code{effect_var}: The affected variable in the relationship.
  \item \code{value}: The value of the target variable in the current iteration.
}
}
\description{
Extract the full MCMC chain associated to the IRFs.
}
