% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bnpgrangertest.R
\name{extract_chain_pdf}
\alias{extract_chain_pdf}
\title{Extract the full MCMC chain associated to the (1d) posterior predictive pdfs.}
\usage{
extract_chain_pdf(object)
}
\arguments{
\item{object}{An instance of class \code{bnpgrangertest}.}
}
\value{
A tibble with the chain of the (1d) posterior predictive pdfs
and the following variables:
\itemize{
  \item \code{iter}: The MCMC iteration.
  \item \code{horizon}: The IRF horizon.
  \item \code{var}: The variable for which the predictive pdf is computed.
  \item \code{value}: The value of the target variable in the current iteration.
}
}
\description{
Extract the full MCMC chain associated to the (1d) posterior predictive pdfs.
}
