% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bnpgrangertest.R
\name{plot_gamma}
\alias{plot_gamma}
\title{Plot the posterior distribution of gamma as a tile plot.}
\usage{
plot_gamma(object)
}
\arguments{
\item{object}{An instance of class \code{bnpgrangertest}.}
}
\value{
A \code{ggplot} object that can be further
customized with the \code{ggplot2} package.
}
\description{
Plot the posterior distribution of gamma as a tile plot.
}
