% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bnpgrangertest.R
\name{plot_irf}
\alias{plot_irf}
\title{Plot the posterior distribution of the IRFs as a collection of line plots.}
\usage{
plot_irf(object)
}
\arguments{
\item{object}{An instance of class \code{bnpgrangertest}.}
}
\value{
A \code{ggplot} object that can be further
customized with the \code{ggplot2} package.
}
\description{
Plot the posterior distribution of the IRFs as a collection of line plots.
}
