% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bnpgrangertest.R
\name{plot_pdf}
\alias{plot_pdf}
\title{Plot the posterior predictive pdf as a collection of density plots.}
\usage{
plot_pdf(object)
}
\arguments{
\item{object}{An instance of class \code{bnpgrangertest}.}
}
\value{
A \code{ggplot} object that can be further
customized with the \code{ggplot2} package.
}
\description{
Plot the posterior predictive pdf as a collection of density plots.
}
