% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bnpgrangertest.R
\name{setup}
\alias{setup}
\title{Check Julia setup}
\usage{
setup()
}
\description{
Check that Julia (v1.0.0+) can be started and install  ANOVADDPTest.jl.
For more information about the setup and discovery of Julia, see
the JuliaConnectoR's package documentation, section "Setup".
}
