% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bnpgrangertest.R
\name{summarize_gamma}
\alias{summarize_gamma}
\title{Summarize the posterior distribution of gamma.}
\usage{
summarize_gamma(object)
}
\arguments{
\item{object}{An instance of class `bnpgrangertest`.}
}
\value{
A tibble with the posterior probability of each hypothesis
(one per row), and the following variables:
\itemize{
  \item `cause_id`: the causing variable in the relationship.
  \item `effect_id`: the affected variable in the relationship.
  \item `prob`: The posterior probability of the causal relationship.
}
}
\description{
Summarize the posterior distribution of gamma.
}
