% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bnpgrangertest.R
\name{summarize_pdf}
\alias{summarize_pdf}
\title{Summarize the posterior predictive distribution.}
\usage{
summarize_pdf(object)
}
\arguments{
\item{object}{An instance of class \code{bnpgrangertest}.}
}
\value{
A tibble with each posterior predictive pdf,
and the following variables:
\itemize{
  \item \code{horizon}: the IRF horizon.
  \item \code{cause_var}: the causing variable in the relationship.
  \item \code{effect_var}: the affected variable in the relationship.
  \item \code{y}: the grid point.
  \item \code{pdf}: The posterior predictive pdf.
}
}
\description{
Summarize the posterior predictive distribution.
}
