% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bnpgrangertest.R
\name{bnpgrangertest}
\alias{bnpgrangertest}
\title{Perform the Granger causality test described in Gutiérrez et al. (2023).}
\usage{
bnpgrangertest(
  Y,
  p = 1L,
  z0 = 1,
  q0 = 1,
  a0p = 1,
  b0p = 1,
  a0s = 8,
  b0s = 4,
  v0 = ncol(Y) + 1L,
  S0 = diag(ncol(Y)),
  warmup = 2000L,
  iter = 4000L,
  thin = 1L,
  hmax = 1L,
  seed = 1L,
  grid_npoints = 50L,
  grid_lbs = apply(Y, 2, function(x) mean(x) - 3 * stats::sd(x)),
  grid_ubs = apply(Y, 2, function(x) mean(x) + 3 * stats::sd(x))
)
}
\arguments{
\item{Y}{A \eqn{T \times N} response matrix with \eqn{T} periods.}

\item{p}{The number of lags in the model.}

\item{z0}{The hyperparameter \eqn{\zeta_{0}}.}

\item{q0}{The hyperparameter \eqn{q_{0} := \sigma_{0}^{2}}.}

\item{a0p}{The hyperparameter \eqn{a_{0p}}.}

\item{b0p}{The hyperparameter \eqn{b_{0p}}.}

\item{a0s}{The hyperparameter \eqn{a_{0s}}.}

\item{b0s}{The hyperparameter \eqn{b_{0s}}.}

\item{v0}{The hyperparameter \eqn{v_{0}}.}

\item{S0}{The hyperparameter \eqn{S_{0}}.}

\item{warmup}{The number of warmup mcmc iterations.}

\item{iter}{The total number of mcmc iterations.}

\item{thin}{The period for saving samples.}

\item{hmax}{The maximum horizon for computing
both the IRFs and the posterior predictive pdfs.}

\item{seed}{The seed for random number generation.}

\item{grid_npoints}{The number of points y0 for computing each posterior
predictive pdf. The final grid for the ith variable is conformed by
`grid_npoints` equispaced points from \code{grid_lbs[i]} to
\code{grid_ubs[i]},  see the arguments \code{grid_lb} and \code{grid_ub}.}

\item{grid_lbs}{A vector of lower bounds, see \code{grid_npoints}.}

\item{grid_ubs}{A vector of upper bounds, see \code{grid_npoints}.}
}
\value{
An object of class \code{bnpgrangertest}.
}
\description{
Perform the Granger causality test described in Gutiérrez et al. (2023).
}
