% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{example_dataset}
\alias{example_dataset}
\title{A simulated dataset for illustration purposes.}
\format{
A matrix with 200 rows and 2 columns where the 1st variable Granger causes
the 2nd variable, but not the other way around.
}
\usage{
example_dataset
}
\description{
A simulated dataset for illustration purposes.
}
\keyword{datasets}
