% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bnpgrangertest.R
\name{summarize_irf}
\alias{summarize_irf}
\title{Summarize the posterior distribution of the IRF.}
\usage{
summarize_irf(object)
}
\arguments{
\item{object}{An instance of class `bnpgrangertest`.}
}
\value{
A tibble with the posterior mean of each requested IRF,
and the following variables:
\itemize{
  \item horizon: the IRF horizon.
  \item cause_id: the causing variable in the relationship.
  \item effect_id: the affected variable in the relationship.
  \item irf: The posterior mean of the IRF.
}
}
\description{
Summarize the posterior distribution of the IRF.
}
