% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bnpgrangertest.R
\name{summarize_pdf}
\alias{summarize_pdf}
\title{Summarize the posterior predictive distribution.}
\usage{
summarize_pdf(x)
}
\arguments{
\item{fit}{An object of class `bnpgrangertest`.}
}
\value{
A tibble with each posterior predictive pdf,
and the following variables:
\itemize{
  \item horizon: the IRF horizon.
  \item cause_id: the causing variable in the relationship.
  \item effect_id: the affected variable in the relationship.
  \item y: the grid point.
  \item pdf: The posterior predictive pdf.
}
}
\description{
Summarize the posterior predictive distribution.
}
